requirements.txt
PyOpenGL
PySide2